splitted operations and field declaration, compiles
