+++ cuda/fft_kernel.cu
@@ -24,6 +24,7 @@
 #include <iostream>
 #include "fft_kernel.h"
 #include "device_field.h"
+#include "device_field_operators.h"
 
 #define LOG_NUM_THREADS 11 
 #define NUM_THREADS 1 << LOG_NUM_THREADS
